+++ src/base/b-dummy/interface.ts
@@ -9,7 +9,9 @@
 import type { ImageLoader, imageLoaderFactory } from 'core/dom/image';
 import type { InViewAdapter } from 'core/dom/in-view';
 import type { ResizeWatcher } from 'core/dom/resize-observer';
+
 import type updateOn from 'core/component/directives/update-on/engines';
+import type iObserveDOM from 'traits/i-observe-dom/i-observe-dom';
 
 export interface Directives {
 	imageFactory: typeof imageLoaderFactory;
@@ -21,4 +23,5 @@ export interface Directives {
 
 export interface Modules {
 	resizeWatcher: typeof ResizeWatcher;
+	iObserveDOM: typeof iObserveDOM;
 }