+++ tests/cases.js
@@ -14,7 +14,7 @@ module.exports = [
 	'--test-entry base/b-virtual-scroll/test',
 	'--test-entry base/b-tree/test',
 	'--test-entry base/b-sidebar/test',
-	'--test-entry base/b-slider/test --browsers chrome --device iPhone_11',
+	'--test-entry base/b-slider/test --device iPhone_11',
 	'--test-entry base/b-list/test',
 	'--test-entry base/b-window/test',
 	'--test-entry base/b-image/test',
@@ -24,6 +24,7 @@ module.exports = [
 	'--test-entry super/i-block/modules/daemons/test',
 	'--test-entry core/dom/image/test',
 	'--test-entry core/dom/in-view/test',
-	'--test-entry core/component/directives/update-on/test',
-	'--test-entry traits/i-lock-page-scroll/test'
+	'--test-entry traits/i-lock-page-scroll/test',
+	'--test-entry traits/i-observe-dom/test',
+	'--test-entry core/component/directives/update-on/test'
 ];